Create real APIs for the CIM repository and separate returns
1. Originally, mocke just accepted the **params from the imethodcall mock
into each of the repository responders but that was very messy for
what is to be a real CIM repository.  We modified the repository API
for each server repository request method handler to use the same
interface as the WBEMConnection client methods.  Then we added code
for each request to Fake_WBEMConnection to map the **params into
specific arguments for each method.  This completely separates the
implementation of the CIM repository from the client so that in the
future we could use the same CIM repository for the basis for
an actual CIM server.

2. All of code that mapped request and response data to be compatible
with the imethodcall dictionary and response tuple was removed from
the CIM repository so it returns the same definition of objects. This
code was used to create corresponding calls for each request method in
Fake_WBEMConneciton to map from the imethod **params to the parameters
required for each request method and to map the method response to a
tuple compatible with imethodcalls

Now the server request handler methods in CIM repository look like what
is in WBEMConnection for the corresponding client methods and return the same definition of objects as WBEMConnection.

All the mapping from and to imethod call is in Fake_WBEMConnection.