Change Name of the file and class CIMRepository to MainProvider.
This set of changes also refactors a number of the methods to remove
verbose code and complete the last steps to making the Provider code
free of the use of CIMClassName on the request side since that is not
really a requirement or even a simplification for the WBEM server
responder methods.

NOTE: There is still one use of CIMClassName; the responses to the
associator/reference operations for classes where the responses
are packaged into CIMClassName as the last step in generating the class
level rsponses.